pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "gdhbot-py"
version = "0.1.0"
[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3"
Flask = "^1.1"
gspread = "^3.6"
oauth2client = "^4.1"
cryptography = "^2.9"
[tool.poetry.dev-dependencies]
